+++ src/main/java/net/frozenblock/wilderwild/entity/AncientHornVibration.java
@@ -93,7 +93,6 @@ import net.minecraft.world.phys.HitResult;
 import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.joml.Vector3f;
 
 //TODO: Fix rendering (Renders too bright or too dark depending on direction)
 
@@ -170,9 +169,9 @@ public class AncientHornVibration extends AbstractArrow {
 					new FloatingSculkBubbleParticleOptions(
 						size,
 						20 + random.nextInt(40),
-						new Vector3f(
+						new Vec3(
 							FloatingSculkBubbleParticleOptions.getRandomVelocity(random, size),
-							0.05F,
+							0.05D,
 							FloatingSculkBubbleParticleOptions.getRandomVelocity(random, size)
 						)
 					),
@@ -215,9 +214,9 @@ public class AncientHornVibration extends AbstractArrow {
 					new FloatingSculkBubbleParticleOptions(
 						0,
 						60,
-						new Vector3f(
+						new Vec3(
 							FloatingSculkBubbleParticleOptions.getRandomVelocity(random, 0),
-							0.05F,
+							0.05D,
 							FloatingSculkBubbleParticleOptions.getRandomVelocity(random, 0)
 						)
 					),